cleanup: move cli script to separate file
